Re-enable failure from conflict of providing libs
Linkerconfig originally failed when there is a conflict of providing
libs from different namespaces, but this was temporarily disabled
because of libz conflict from system and vendor namespace. Based on the
work from b/313806237, libz is now only offered from system namespace,
so there is no more library conflict. This change is to re-enable
failure when there is any conflict.

Bug: 297821005
Test: Cuttlefish boot succeeded
Change-Id: I8cf0a52ee36bff5ee15b70017441a0c9fe26f859