Link sanitizer LLNDK libs
Vendor/product apexes need to link sanitizer LLNDK libraries from
system. For now, vendor/product apex requires :sanitizer. But with VNDK
deprecation the requirement is missing, and vendor apex installation
fails. (Note that linkerconfig is called with --strict mode on apex
installation, which aborts when a requirement is not met)

Previously, sanitizer VNDK libs are linked from the vndk namespace and
sanitizer LLNDK libs are linked from the system namespace with
:sanitizer alias.

Now, sanitizer LLNDK is linked to vendor/product apexes.

Bug: 335491075
Bug: 329366340
Test: atest --test-mapping .
Test: atest VendorApexHostTestCases
Change-Id: I8fe57ffc21dc7cc3ea03f92a9e5cc1948b2d8c77